refactor(plugin-file-manager): separate MIME matching rule
