+++ packages/plugins/@nocobase/plugin-file-manager/src/server/rules/index.ts
@@ -7,12 +7,7 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-export { default as mimetype } from './mimetype';
-export {
-  ACTIVE_CONTENT_EXTENSIONS,
-  ACTIVE_CONTENT_MIMETYPES,
-  isDisallowedActiveContent,
-  matchesMimePattern,
-} from './active-content';
+export { default as mimetype, matchesMimePattern } from './mimetype';
+export { ACTIVE_CONTENT_EXTENSIONS, ACTIVE_CONTENT_MIMETYPES, isDisallowedActiveContent } from './active-content';
 
 // TODO(feature): 提供注册新规则的方法，规则可动态添加
