+++ packages/plugins/@nocobase/plugin-file-manager/src/server/rules/mimetype.ts
@@ -7,7 +7,24 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { matchesMimePattern } from './active-content';
+import match from 'mime-match';
+
+export function normalizeMimePattern(pattern: string | string[] = '*') {
+  const value = Array.isArray(pattern) ? pattern.join(',') : pattern;
+  return value
+    .toString()
+    .split(',')
+    .map((item) => item.trim().toLowerCase())
+    .filter(Boolean);
+}
+
+export function matchesMimePattern(mimetype: string, pattern: string | string[] = '*') {
+  const normalizedPattern = normalizeMimePattern(pattern);
+  if (!normalizedPattern.length || normalizedPattern.includes('*')) {
+    return true;
+  }
+  return normalizedPattern.some(match(mimetype.toLowerCase()));
+}
 
 export default function (file, options: string | string[] = '*'): boolean {
   return matchesMimePattern(file.mimetype, options);
